+++ pegainfer-qwen35/src/forward/batch_decode.rs
@@ -11,18 +11,18 @@ use pegainfer_core::kv_pool::KvState;
 use pegainfer_core::tensor::HiddenStates;
 use pegainfer_frontend::sampler::SamplingParams;
 
-use super::batch_decode_graph::BATCH_BUCKETS;
-use super::batch_decode_graph::BatchDecodeGraphState;
-use super::batch_decode_graph::bucket_for;
-use super::decode_buffers::BatchDecodeBuffers35;
-use super::recurrent_state::LinearStatePointerTables;
-use super::recurrent_state::RecurrentState;
-use super::weights::FullAttentionLayer;
-use super::weights::LayerKind;
-use super::weights::LinearAttentionLayer;
-use super::weights::Qwen35Model;
-use super::weights::TransformerBlock35;
+use crate::decode_buffers::BatchDecodeBuffers35;
+use crate::forward::batch_decode_graph::BATCH_BUCKETS;
+use crate::forward::batch_decode_graph::BatchDecodeGraphState;
+use crate::forward::batch_decode_graph::bucket_for;
 use crate::ops;
+use crate::recurrent_state::LinearStatePointerTables;
+use crate::recurrent_state::RecurrentState;
+use crate::weights::FullAttentionLayer;
+use crate::weights::LayerKind;
+use crate::weights::LinearAttentionLayer;
+use crate::weights::Qwen35Model;
+use crate::weights::TransformerBlock35;
 
 static LOG_UNCOMPILED_DECODE_ROUTE: std::sync::Once = std::sync::Once::new();
