+++ pegainfer-qwen35/src/scheduler.rs
@@ -66,12 +66,12 @@ use self::telemetry::*;
 use self::tp::*;
 use crate::Qwen35DecodeOverlap;
 use crate::Qwen35SchedulerPolicy;
-use crate::batch_decode_graph::BatchDecodeGraphState;
 use crate::executor::DecodeRequestResult;
 use crate::executor::DecodeResult;
 use crate::executor::PrefillRequestResult;
 use crate::executor::PrefillResult;
 use crate::executor::RequestId;
+use crate::forward::batch_decode_graph::BatchDecodeGraphState;
 use crate::logprobs::snapshot_requested_logprobs;
 use crate::recurrent_state::RecurrentState;
 use crate::tp_executor::DropExpectation;
