+++ pegainfer-qwen35/src/tp_executor.rs
@@ -31,8 +31,8 @@ use crate::executor::PrefillRequestResult;
 use crate::executor::PrefillResult;
 use crate::executor::PrefillStepItem;
 use crate::executor::RequestId;
+use crate::forward::prefill::PREFILL_CHUNK_LEN;
 use crate::logprobs::snapshot_requested_logprobs;
-use crate::prefill::PREFILL_CHUNK_LEN;
 use crate::prefill_buffers::GdrChunkwiseScratch35;
 use crate::recurrent_state::LinearStatePointerTables;
 use crate::recurrent_state::RecurrentState;
